Fix search bar showing up on details page, Fix details page not re-rendering on page refresh.
